refactor(transform): isolate shaker binding provenance
